run script file feature

+++ src/commands/alias.h
@@ -8,7 +8,7 @@
 
 int alias(std::string i){
     // add command
-    int found = i.find("=");
+    size_t found = i.find("=");
     // std::string cmd_name = i.substr(0, found);
     // std::string cmd_content = i.substr(found+1, i.size()-1);
     // alias_map.insert(cmd_name, cmd_content);
